+++ pyinterpolate/kriging/models/point/simple_kriging.py
@@ -26,8 +26,7 @@ def simple_kriging(
         neighbors_range=None,
         no_neighbors=1,
         use_all_neighbors_in_range=False,
-        allow_approximate_solutions=False,
-        err_to_nan=False
+        allow_approximate_solutions=False
 ) -> List:
     """
     Function predicts value at unknown location with Ordinary Kriging technique.
@@ -59,7 +58,7 @@ def simple_kriging(
         ``True``: if the real number of neighbors within the ``neighbors_range`` is greater than the
         ``number_of_neighbors`` parameter then take all of them anyway.
 
-    allow_approx_solutions : bool, default=False
+    allow_approximate_solutions : bool, default=False
         Allows the approximation of kriging weights based on the OLS algorithm. We don't recommend set it to ``True``
         if you don't know what are you doing. This parameter can be useful when you have clusters in your dataset,
         that can lead to singular or near-singular matrix creation.